src/search/requirements.txt
qdrant-client
rank-bm25
fastapi
uvicorn[standard]
httpx
sentence-transformers
mecab-python3
